Moved fs to it's own crate, build failing due to cyclic dependency on rope
